requirements.txt
langchain-community
langchain-ollama
langchain-google-genai
langgraph
langgraph-checkpoint-redis
langchain-openai
langchain-anthropic
langchain-deepseek
openpyxl
fastapi
uvicorn
asyncpg
redis
aioredis
sqlalchemy[asyncio]
psycopg[binary,pool]